+++ whatsNew.md
@@ -1,3 +1,4 @@
+ * Played for two clubs - Nov 28, 2015
  * Player Scoring Sequences (plotly version) - Nov 21, 2015
  * Player Scoring Sequences - November 21, 2015
  * Added Player Comparison chart - November 13, 2015